requirements.txt
fastapi == 0.95.1
sqlmodel == 0.0.8
python-decouple == 3.8
uvicorn == 0.22.0
passlib == 1.7.4
python-jose == 3.3.0
python-multipart == 0.0.6
bcrypt == 4.0.1
httpx == 0.24.0
pytest == 7.3.1
gunicorn == 20.1.0
pyodbc == 4.0.32
python-decouple == 3.8
python-jose == 3.3.0
python-dotenv == 1.0.0
